+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Belarusian translation (thanks, Nelly Simkova!)
+
 ## [0.17] - 2018-07-08
 ### Packaging changes
 - New dependency: [requests](https://pypi.org/project/requests/)